Some rearranging of files
