+++ api/src/adapters/search/__init__.py
@@ -1,0 +1,4 @@
+from src.adapters.search.opensearch_client import SearchClient, get_opensearch_client
+from src.adapters.search.opensearch_config import get_opensearch_config
+
+__all__ = ["SearchClient", "get_opensearch_client", "get_opensearch_config"]
